refactor: change MsgOfInterest tags to Hashset
This is a minor enhancement/refactoring:
- MsgOfInterest tags: replace string to HashSet
- rename "websocketcodec" file to "websocket" (align with convention)
